Refactor code to use signed data types.
This is required to handle diffing between files or between
individual timestamps. Since we use 64bit data types, we
can still handle a lot of allocations this way (a bit more than
9 exabytes).

+++ accumulatedtracedata.cpp
@@ -289,7 +289,7 @@ bool AccumulatedTraceData::read(istream& in)
 
             const auto& info = allocationInfos[allocationInfoIndex.index];
             auto& allocation = findAllocation(info.traceIndex);
-            if (!allocation.allocations || allocation.leaked < info.size) {
+            if (!allocation.allocations || static_cast<uint64_t>(allocation.leaked) < info.size) {
                 if (!fromAttached) {
                     cerr << "inconsistent allocation info, underflowed allocations of " << info.traceIndex << endl;
                 }
